+++ crates/flux-sysroot/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "flux-sysroot"
+version = "0.1.0"
+
+edition.workspace = true
+
+[dependencies]
+home.workspace = true
+
+[lints]
+workspace = true
