+++ flux-dev/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "flux-dev"
+version = "0.1.0"
+
+edition.workspace = true
+
+[lints]
+workspace = true
