+++ tests/Cargo.toml
@@ -8,6 +8,10 @@ edition.workspace = true
 doctest = false
 test = false
 
+[dependencies]
+flux-dev.workspace = true
+flux-sysroot.workspace = true
+
 [dev-dependencies]
 itertools.workspace = true
